Add navigation to ServingRuntime and set plugin metadata in ConsoleRemotePlugin
These changes add a new top-level section in the left main navigation of OpenShift console. The new navigation is going to be below `Workloads`. Inside of it, a navigation item for Serving Runtimes is added. At the moment, OpenShift default list and details pages are used.

Additionally, the plugin metadata is now being specified in `ConsoleRemotePlugin` (in webpack.config.ts):
* The console-extensions.json is no longer used and, instead, the extensions are configured in the new console-extensions.ts file.
* The `consolePlugin` is also no longer used and, instead, the metadata is specified directly `ConsoleRemotePlugin` in webpack.config.ts file.