Add Deployment Mode and Model URI to details of ISVCs
Both fields are added to the right column of the details page of InferenceServices.

The _deployment mode_ is taken from the `status` of the InferenceService rather than the annotation, asusming this is the mode in effect.

The _Model URI_ is shown from the spec.predictor.model.storageUri field of the ISVC.