Add a custom page for creating InferenceServices
The custom page is in CreateInferenceService.tsx. The console-extensions.ts and webpack.config.ts files were updated accordingly. The custom page is incomplete because it contains markup related to a sample page from the console-template-plugin. This custom page is supposed to have a form to allow creating an InferenceService more easily than with YAML.